pyproject.toml
[tool.poetry]
name = "pyes"
version = "1.1.3"
description = ""
authors = ["Lorenzo Castellino <castellino96@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pyinstaller = "~5"
ppg = {git = "ssh://git@github.com/Kastakin/ppg.git", rev = "main"}
pandas = "^1.5.0"
openpyxl = "^3.0.9"
pyqtgraph = "^0.13.0"
pyside6-essentials = "~6.4"
libeq = {git = "https://github.com/Kastakin/libeq.git"}

[tool.poetry.group.dev.dependencies]
tbump = "^6.7.0"
isort = "^5.10.1"
black = "^22.3.0"
pre-commit = "^2.18.1"
matplotlib = "^3.6.0"
scipy = "^1.9.1"
pyinstrument = "^4.3.0"
Pillow = "^10.0.0"
ruff = "~0.0"
pytest = "^7.2.0"
pytest-sugar = "^0.9.6"
pytest-cov = "^4.0.0"
pytest-qt = "^4.2.0"
pytest-xvfb = "^2.0.0"
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
pythonpath = "src/main/python/pyes"
addopts = "--cov=src/main/python/pyes --cov-report=html --cov-config=.coveragerc"
qt_api = "pyside6"

[tool.ruff.format]
# Enable preview style formatting.
preview = true